# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/accounts/add_account": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Adds 1 account.",
        "description": "Adds the account in the request.",
        "operationId": "addAccount",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns posted account",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/accounts/edit_account/{id}": {
      "put": {
        "tags": [
          "Accounts"
        ],
        "summary": "Edits account.",
        "description": "Replaces account information with information from the request.",
        "operationId": "editAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns edit account",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/accounts/delete_account/{id}": {
      "delete": {
        "tags": [
          "Accounts"
        ],
        "summary": "Deletes account.",
        "description": "Deletes account with specified Id.",
        "operationId": "deleteAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns deleted account",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Account": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string",
          "example": "TestUserName"
        },
        "password": {
          "type": "string",
          "example": "123456"
        },
        "firstName": {
          "type": "string",
          "example": "Joe"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "address": {
          "type": "string",
          "example": "12 Test Street, ID, USA"
        },
        "email": {
          "type": "string",
          "example": "Joe_Doe@gmail.com"
        }
      }
    },
...
  }
}